Update HLSL_ShaderModel6_5.md
Hi,

Before my change, it says "Note how subset with mask.x == 0x0b refers to lane 1, which is either inactive or is a helper lane". Since mask 0x0b is for lane 0 and it's Prefix, lane 1 shouldn't be considered anyway. I think set the second bit of mask in lane 2 will reduce possibility of confusion.
It also mentioned "The groups are assumed to be non-intersecting". However, mask 0xb and 0x9 have overlaps in the example, so I suggest to update 0xb to 0x2.
BTW, do you really want the groups to be non-intersecting? I think there are quite some scenarios in real world needs to reuse same lane in different groups.

Thanks,
Mengbo