Update VariableRateShading.md
Tried to fix the following sentence to make it match the tables below (for some interpretations).  It may actually be better to delete the entire sentence:

> The mask consists of the of coverages from pixels from left-to-right, then top-to-bottom (column-major) order.

First, it's definitely row-major order, not column-major. Each row of fine pixels are contiguous in the sample mask, not each column.

Second, when evaluating an attribute with centroid evaluation, the definition states to use the "first covered sample". The "first" means the lowest-order bit that is set.  I interpret that to mean that the "order" of the sample mask should be read as lowest-order to highest-order.  Given the tables in this diagram, I would say the order is "right-to-left, then bottom-to-top (row-major)", hence this change.