Remove nonworking code from Dynamic Resources spec
Passing a value from a resource array to an intrinsic doesn't work in
DXC. That will be addressed at some point, but in the meantime, we don't
want the leading people to use nonfunctional code.

This updates the spec text to avoid the broken usage and makes the
example more illustrative of how the resource heaps work.

See https://github.com/microsoft/DirectXShaderCompiler/issues/4520